+++ service/src/test/java/com/epam/esm/service/converter/impl/GiftCertificateCreateDtoToGiftCertificateConverterTest.java
@@ -5,6 +5,8 @@ import com.epam.esm.model.entity.GiftCertificate;
 import com.epam.esm.service.converter.Converter;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -12,7 +14,7 @@ public class GiftCertificateCreateDtoToGiftCertificateConverterTest {
 	private static final String NAME = "name";
 	private static final String DESCRIPTION = "description";
 	private static final int DURATION = 1;
-	private static final double PRICE = 2.3;
+	private static final BigDecimal PRICE = BigDecimal.valueOf(2.3);
 
 	@Test
 	public void convertShouldReturnEntityWhenPassedCorrectDto() {
