+++ service/src/test/java/com/epam/esm/service/merger/impl/GiftCertificateUpdateDtoIntoGiftCertificateMergerTest.java
@@ -5,6 +5,7 @@ import com.epam.esm.model.entity.GiftCertificate;
 import com.epam.esm.service.merger.Merger;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -14,7 +15,7 @@ public class GiftCertificateUpdateDtoIntoGiftCertificateMergerTest {
 	private static final int ENTITY_ID = 1;
 	private static final String ENTITY_NAME = "entity_name";
 	private static final String ENTITY_DESC = "entity_desc";
-	private static final double ENTITY_PRICE = 1.2;
+	private static final BigDecimal ENTITY_PRICE = BigDecimal.valueOf(1.2);
 	private static final int ENTITY_DURATION = 2;
 	private static final String DTO_DESC = "dto_desc";
 	private static final LocalDateTime ENTITY_CREATE_DATE = null;
